Hacking in a few IP Options.
Changed congestionLevel from int to double.
Added tags for IP Option RouteRecord and Strict source routing.
Trying to hack-in RouteRecord on all TCP-Data and Strict Source routing
on Acks. (Need to propagate the RouteRecord to TCP and copy it to the
Ack as Strict source).

+++ src/inet/routing/dmpr/DmprInterfaceData.h
@@ -24,15 +24,15 @@ namespace inet {
 class INET_API DmprInterfaceData : public InterfaceProtocolData
 {
   private:
-    int congestionLevel;
+    double congestionLevel;
 
   public:
     DmprInterfaceData();
     virtual ~DmprInterfaceData();
     virtual std::string str() const override;
     virtual std::string detailedInfo() const override;
-    int getCongestionLevel() const;
-    void setCongestionLevel(int congestionLevel);
+    double getCongestionLevel() const;
+    void setCongestionLevel(double congestionLevel);
     void incCongestionLevel();
